Move functionality in rime_lib.py.
Consolidate RIME code in rime_lib.py.
Place RIME example code in rime.py.
